requirements.txt
Django>=4.2
django-allauth
google-api-python-client
google-auth
gunicorn 